src/app/core/interceptor/interceptor.service.ts
import { Injectable } from '@angular/core';
import {
  HttpInterceptor,
  HttpRequest,
  HttpHandler,
  HttpEvent,
} from '@angular/common/http';
import { Observable } from 'rxjs';
import { environment } from 'src/environments/environment.development';

@Injectable()
export class InterceptorService implements HttpInterceptor {
  private apiKey = environment.apiKey;

  intercept(
    request: HttpRequest<any>,
    next: HttpHandler
  ): Observable<HttpEvent<any>> {
    const modifiedRequest = request.clone({
      setParams: {
        apikey: this.apiKey,
      },
    });
    return next.handle(modifiedRequest);
  }
}
